src/app/core/interceptors/auth.interceptor.ts
import {
  HttpEvent, HttpHandler, HttpInterceptor, HttpRequest
} from '@angular/common/http';
import { Injectable } from '@angular/core';
import { TokenService } from '@auth/services/token.service';
import { Observable } from 'rxjs';

@Injectable()
export class AuthInterceptor implements HttpInterceptor {
  constructor(private tokenService: TokenService) {}

  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
    // eslint-disable-next-line @typescript-eslint/no-unused-vars
    const accessToken = this.tokenService.getAccessToken();
    return next.handle(request);
  }
}
